# agenthub-cli/docs/openapi-baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agenthub",
    "description": "Home Agent Hub. Approved shared context, daily worklogs, and alignment snapshots. Authenticated REST and MCP. Not an LLM.",
    "version": "1.3.0"
  },
  "paths": {
...
    "/api/v1/grants": {
      "post": {
        "summary": "Put Grant",
        "operationId": "put_grant_api_v1_grants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/workspaces/me": {
      "get": {
        "summary": "My Workspace",
        "operationId": "my_workspace_api_v1_workspaces_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/workspaces/me/nodes": {
      "post": {
        "summary": "Create My Node",
        "operationId": "create_my_node_api_v1_workspaces_me_nodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/workspaces/me/uploads": {
      "post": {
        "summary": "Upload My Node",
        "operationId": "upload_my_node_api_v1_workspaces_me_uploads_post",
        "parameters": [
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Id"
            }
          },
          {
            "name": "admin_reason",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Admin Reason"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_my_node_api_v1_workspaces_me_uploads_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/publish-requests": {
      "get": {
        "summary": "Get Publish List",
        "operationId": "get_publish_list_api_v1_publish_requests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Post Publish",
        "operationId": "post_publish_api_v1_publish_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/publish-requests/{request_id}/approve": {
      "post": {
        "summary": "Approve Publish",
        "operationId": "approve_publish_api_v1_publish_requests__request_id__approve_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/publish/plans": {
      "post": {
        "summary": "Post Plan",
        "operationId": "post_plan_api_v1_publish_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v1/agents/me/heartbeat": {
      "post": {
        "summary": "Heartbeat",
        "operationId": "heartbeat_v1_agents_me_heartbeat_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Heartbeat V1 Agents Me Heartbeat Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/console/publish": {
      "get": {
        "summary": "Console Publish",
        "operationId": "console_publish_console_publish_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Console Publish",
        "operationId": "console_publish_console_publish_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/console/publish/requests": {
      "post": {
        "summary": "Console Publish Request",
        "operationId": "console_publish_request_console_publish_requests_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_my_node_api_v1_workspaces_me_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_my_node_api_v1_workspaces_me_uploads_post"
      },
...
      "GrantIn": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": ""
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "project_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Project Ids"
          },
          "scopes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scopes"
          },
          "share_classes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Classes"
          },
          "revoked": {
            "type": "boolean",
            "title": "Revoked",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "agent_id"
        ],
        "title": "GrantIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeartbeatIn": {
        "properties": {
          "status": {
            "type": "string",
            "maxLength": 32,
            "title": "Status",
            "default": "idle"
          }
        },
        "type": "object",
        "title": "HeartbeatIn"
      },
...
      "NodeIn": {
        "properties": {
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "file"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type",
            "default": "text/plain"
          },
          "admin_reason": {
            "type": "string",
            "title": "Admin Reason",
            "default": ""
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "owner_agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Agent Id"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "NodeIn"
      },
...
      "PlanIn": {
        "properties": {
          "prefix": {
            "type": "string",
            "title": "Prefix",
            "default": ""
          },
          "repo": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Repo"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "create"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility",
            "default": "public"
          },
          "license": {
            "type": "string",
            "title": "License",
            "default": "MIT"
          },
          "copyright_holder": {
            "type": "string",
            "title": "Copyright Holder",
            "default": "Xiwei Chen"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "repo"
        ],
        "title": "PlanIn"
      },
...
      "PublishIn": {
        "properties": {
          "node_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Node Ids"
          },
          "target_owner": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Target Owner"
          },
          "repo": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Repo"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "default": "main"
          },
          "create_repo": {
            "type": "boolean",
            "title": "Create Repo",
            "default": false
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          }
        },
        "type": "object",
        "required": [
          "target_owner",
          "repo"
        ],
        "title": "PublishIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}